backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart>=0.0.6

# LLM Providers
openai>=1.99.5
anthropic==0.39.0
google-generativeai==0.8.3

# HTTP Client
httpx>=0.28.1,<1.0.0

# Security
cryptography>=41.0.0

# Caching
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# Core Dependencies (resolved for compatibility)
tiktoken>=0.7.0
typing-extensions>=4.8.0
